first optimize: seperate the hooks from the pages
